Renamed sample maze file
