+++ src/components/SampleMazes.vue
